% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA.R
\name{ancestr}
\alias{ancestr}
\title{Hidden Ancestries function}
\usage{
ancestr(D = NULL, k = 0, t = 0, x_0 = NULL)
}
\arguments{
\item{D}{dataframe: This dataframe should be in the format described in our data formatting document.}

\item{k}{integer: This parameter is the number of reference ancestries present in your dataframe D.}

\item{t}{integer: variable t is the number of the observed allele frequency
column the user wishes to run through SQP. If not specified, the function will just use the first
observed vector.}

\item{x_0}{: x_0 is the starting guess for the SLSQP algorithm.}
}
\value{
Estimated ancestry proportions
}
\description{
Estimates ancestry proportions in heterogeneous allele frequency data
}
\examples{
data("ancestryData")
ancestr(ancestryData, 5, 1)

}
\author{
Gregory Matesi, \email{gregory.matesi@ucdenver.edu}
}
\keyword{genomics}
