% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryData-data.R
\docType{data}
\name{ancestryData}
\alias{ancestryData}
\title{Example allele frequency data}
\format{SNP, chromosome, refernce allele frequencies, observed allele frequencies}
\usage{
data(ancestryData)
}
\description{
reference data is 1000 Genomes and NAM. 
Observed data is from gnomAD:
african/african american, 
american/admixed latinx, 
and other
}
\examples{
data("ancestryData")
ref <-ancestryData[,c("ref_afr", "ref_eas", "ref_eur", "ref_nam", "ref_sas")]
obs <- cbind(ancestryData$gnomad_afr)
ancestr(ref,obs)
}
\keyword{datasets}
