% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA.R
\name{ancestr}
\alias{ancestr}
\title{Hidden Ancestries function}
\usage{
ancestr(refmatrix, obsvector)
}
\arguments{
\item{refmatrix}{matrix of reference allele frequency data}

\item{obsvector}{observed heterogeneous allele frequency vector}
}
\value{
Estimated ancestry proportions
}
\description{
Estimates ancestry proportions in heterogeneous allele frequency data
}
\examples{
data <- data("packagedata")
ref <-data[,c("ref_afr", "ref_eas", "ref_eur", "ref_nam", "ref_sas")]
obs <- cbind(data$gnomad_afr)
ancestr(ref,obs)

}
\author{
Gregory Matesi, \email{gregory.matesi@ucdenver.edu}
}
\keyword{genomics}
