% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAF.R
\name{updateAF}
\alias{updateAF}
\title{update allele frequencies function}
\usage{
updateAF(D = NULL, ancestry = NULL)
}
\arguments{
\item{D}{is a dataframe: This dataframe must be in the same format as described
in the data format document.}

\item{ancestry}{dataframe: This dataframe is described below in the code.
It is also described in the README.}
}
\value{
data.frame D with an additional column at the end containing updated allele frequencies.
}
\description{
updates allele frequency data in heterogeneous genetic variant data
}
\examples{
write code

}
\author{
Gregory Matesi, \email{gregory.matesi@ucdenver.edu}
}
\keyword{genomics}
